apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

youtube-transcript-api==0.6.2
yt-dlp==2025.1.12

srt==3.5.3

# Database
sqlalchemy==2.0.25
sqlmodel==0.0.24
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
bcrypt==4.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
